python/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pillow==11.1.0
psutil==7.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
mistralai
ipykernel
python-dotenv